Fix the naming issue that appeared from previous merge
It seems after previous mege paths to 3d models are changed from
e.g., "iit-coman" to "coman_urdf". I suspect this is due to the
fact that IIT friends use coman ROS package.

Furthermore, I had to revet the protocol from "package" to "model"
which is compatable with Orocos design.